bounty agora fica com o dinheiro do user q o criou e tanto o dinheiro da bounty como do user passou a ser inteiro

+++ app/src/main/java/com/example/bountynet/Objects/Bounty.kt
@@ -2,7 +2,7 @@ package com.example.bountynet.Objects
 
 data class Bounty(
     val name: String = "Unnamed Bounty",
-    val reward: Double = 0.0,
+    val reward: Int = 0,
     val concluida: Boolean = false,
     val tempo: String = "Unfinished",
     val hunter: String = "None",
